requirements.txt
# Code Function: Final Project
# Date: 2025/05/27, created by: 蕭智強

antiorm==1.2.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cors==1.0.1
db==0.1.1
db-sqlite3==0.0.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.5.1
google-auth==2.40.1
google-auth-oauthlib==1.2.2
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.4.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.4.0
six==1.17.0
SQLAlchemy==2.0.40
tldextract==5.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
